requirements.txt
fastapi
uvicorn
pillow
pdf2image
ollama
pydantic
groq
python-dotenv